Extracted duplicate content into parent
